moved CardProxy into 'app' module
- moved CardProxy to `net.demilich.metastone.gui.cards`
- moved the loadCard() method impl into CardCatalogue
- updated TestBase to load cards by calling CardCatalogue.loadCards()
rather than invoking CardProxy
